Inline remote translation logic into the model
Move the logic into the model so we can drop RemoteTranslations::Caller.
