+++ spec/models/remote_translation_spec.rb
@@ -86,14 +86,202 @@ describe RemoteTranslation, :remote_translations do
 
       expect(Delayed::Job.count).to eq 1
     end
+  end
 
-    it "uses the same remote translations caller and client every time" do
-      client_class = RemoteTranslations::Client
+  describe "#translate_remotely!" do
+    let(:client) { RemoteTranslations::Client }
 
-      expect_any_instance_of(client_class).to receive(:call).and_return([])
+    context "Debates" do
+      let(:debate) { create(:debate) }
 
-      remote_translation.enqueue_remote_translation
-      remote_translation.enqueue_remote_translation
+      it "returns the resource with new translation persisted" do
+        response = ["Título traducido", "Descripción traducida"]
+        expect_any_instance_of(client).to receive(:call).and_return(response)
+
+        create(:remote_translation, remote_translatable: debate, locale: :es)
+
+        expect(debate.translations.count).to eq(2)
+      end
+
+      it "when new translation locale is distinct to default_locale skip length validations" do
+        response = ["TT", "Descripción traducida"]
+        expect_any_instance_of(client).to receive(:call).and_return(response)
+
+        remote_translation = create(:remote_translation, remote_translatable: debate, locale: :es)
+
+        expect(remote_translation.error_message).to be nil
+        expect(debate.translations.count).to eq(2)
+        expect(debate.valid?).to be true
+      end
+
+      it "when new translation locale is distinct to default_locale not skip presence validations" do
+        response = ["", "Descripción traducida"]
+        expect_any_instance_of(client).to receive(:call).and_return(response)
+
+        remote_translation = create(:remote_translation, remote_translatable: debate, locale: :es)
+
+        expect(remote_translation.error_message).to include("can't be blank")
+        expect(debate.translations.count).to eq(1)
+        expect(debate.valid?).to be false
+      end
+
+      it "destroy remote translation instance" do
+        response = ["Título traducido", "Descripción traducida"]
+        expect_any_instance_of(client).to receive(:call).and_return(response)
+
+        create(:remote_translation, remote_translatable: debate, locale: :es)
+
+        expect(RemoteTranslation.count).to eq(0)
+      end
+    end
+
+    context "Proposals" do
+      let!(:proposal) { create(:proposal) }
+
+      it "returns the resource with new translation persisted" do
+        response = ["Título traducido", "Descripción traducida", "Pregunta traducida",
+                    "Resumen traducido", nil]
+        expect_any_instance_of(client).to receive(:call).and_return(response)
+
+        create(:remote_translation, remote_translatable: proposal, locale: :es)
+
+        expect(proposal.translations.count).to eq(2)
+      end
+
+      it "when new translation locale is distinct to default_locale skip lenght validations" do
+        response = ["TT", "Descripción traducida", "Pregunta traducida", "Resumen traducido", nil]
+        expect_any_instance_of(client).to receive(:call).and_return(response)
+
+        remote_translation = create(:remote_translation, remote_translatable: proposal, locale: :es)
+
+        expect(remote_translation.error_message).to be nil
+        expect(proposal.translations.count).to eq(2)
+        expect(proposal.valid?).to be true
+      end
+
+      it "when new translation locale is distinct to default_locale do not skip presence validations" do
+        response = ["", "Descripción traducida", "Pregunta traducida", "Resumen traducido", nil]
+        expect_any_instance_of(client).to receive(:call).and_return(response)
+
+        remote_translation = create(:remote_translation, remote_translatable: proposal, locale: :es)
+
+        expect(remote_translation.error_message).to include("can't be blank")
+        expect(proposal.translations.count).to eq(1)
+        expect(proposal.valid?).to be false
+      end
+
+      it "destroy remote translation instance" do
+        response = ["Título traducido", "Descripción traducida", "Pregunta traducida",
+                    "Resumen traducido", nil]
+        expect_any_instance_of(client).to receive(:call).and_return(response)
+
+        create(:remote_translation, remote_translatable: proposal, locale: :es)
+
+        expect(RemoteTranslation.count).to eq(0)
+      end
+    end
+
+    context "Budget Investments" do
+      let(:budget_investment) { create(:budget_investment) }
+
+      it "returns the resource with new translation persisted" do
+        response = ["Título traducido", "Descripción traducida"]
+        expect_any_instance_of(client).to receive(:call).and_return(response)
+
+        create(:remote_translation, remote_translatable: budget_investment, locale: :es)
+
+        expect(budget_investment.translations.count).to eq(2)
+      end
+
+      it "when new translation locale is distinct to default_locale skip lenght validations" do
+        response = ["TT", "Descripción traducida"]
+        expect_any_instance_of(client).to receive(:call).and_return(response)
+
+        remote_translation = create(:remote_translation, remote_translatable: budget_investment, locale: :es)
+
+        expect(remote_translation.error_message).to be nil
+        expect(budget_investment.translations.count).to eq(2)
+        expect(budget_investment.valid?).to be true
+      end
+
+      it "when new translation locale is distinct to default_locale not skip presence validations" do
+        response = ["", "Descripción traducida"]
+        expect_any_instance_of(client).to receive(:call).and_return(response)
+
+        remote_translation = create(:remote_translation, remote_translatable: budget_investment, locale: :es)
+
+        expect(remote_translation.error_message).to include("can't be blank")
+        expect(budget_investment.translations.count).to eq(1)
+        expect(budget_investment.valid?).to be false
+      end
+
+      it "destroy remote translation instance" do
+        response = ["Título traducido", "Descripción traducida"]
+        expect_any_instance_of(client).to receive(:call).and_return(response)
+
+        create(:remote_translation, remote_translatable: budget_investment, locale: :es)
+
+        expect(RemoteTranslation.count).to eq(0)
+      end
+    end
+
+    context "Comments" do
+      let(:comment) { create(:comment) }
+
+      it "returns the resource with new translation persisted" do
+        response = ["Body traducido"]
+        expect_any_instance_of(client).to receive(:call).and_return(response)
+
+        create(:remote_translation, remote_translatable: comment, locale: :es)
+
+        expect(comment.translations.count).to eq(2)
+      end
+
+      it "when new translation locale is distinct to default_locale skip lenght validations" do
+        response = ["BT"]
+        expect_any_instance_of(client).to receive(:call).and_return(response)
+
+        remote_translation = create(:remote_translation, remote_translatable: comment, locale: :es)
+
+        expect(remote_translation.error_message).to be nil
+        expect(comment.translations.count).to eq(2)
+        expect(comment.valid?).to be true
+      end
+
+      it "when new translation locale is distinct to default_locale not skip presence validations" do
+        response = [""]
+        expect_any_instance_of(client).to receive(:call).and_return(response)
+
+        remote_translation = create(:remote_translation, remote_translatable: comment, locale: :es)
+
+        expect(remote_translation.error_message).to include("can't be blank")
+        expect(comment.translations.count).to eq(1)
+        expect(comment.valid?).to be false
+      end
+
+      it "destroy remote translation instance" do
+        response = ["Body traducido"]
+        expect_any_instance_of(client).to receive(:call).and_return(response)
+
+        create(:remote_translation, remote_translatable: comment, locale: :es)
+
+        expect(RemoteTranslation.count).to eq(0)
+      end
+    end
+
+    describe "#fields_values" do
+      let!(:proposal) { create(:proposal, description: "&Sigma; with sample text") }
+
+      it "sanitizes field values when they contain entity references such as &Sigma;" do
+        field_values_sanitized = [proposal.title, "Σ with sample text", proposal.summary,
+                                  proposal.retired_reason]
+        fake_response = ["translated title", "translated description", "translated summary", nil]
+
+        expect_any_instance_of(client).to receive(:call).with(field_values_sanitized, "es")
+                                                        .and_return(fake_response)
+
+        create(:remote_translation, remote_translatable: proposal, locale: :es)
+      end
     end
   end
 end
